backend/app/services/calendar_sync.py
"""
Calendar Integration Service
"""
from typing import Dict, Any, List, Optional
from datetime import datetime, timedelta
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
from google.auth.transport.requests import Request
from msal import ConfidentialClientApplication
from app.core.config import settings
from app.core.supabase import supabase_client
import structlog

logger = structlog.get_logger()


class CalendarSyncService:
    """Service for syncing with external calendars (Google, Outlook)"""

    def __init__(self, user_id: str):
        self.user_id = user_id

    async def connect_google_calendar(
        self, authorization_code: str
    ) -> Dict[str, Any]:
        """
        Connect user's Google Calendar using OAuth code
        """
        try:
            # TODO: Implement full OAuth flow with google-auth-oauthlib
            # This is a placeholder for the OAuth process

            # For now, store placeholder tokens
            integration_data = {
                "user_id": self.user_id,
                "provider": "google",
                "access_token": "placeholder_access_token",
                "refresh_token": "placeholder_refresh_token",
                "token_expires_at": (
                    datetime.utcnow() + timedelta(hours=1)
                ).isoformat(),
                "sync_enabled": True,
                "created_at": datetime.utcnow().isoformat(),
                "updated_at": datetime.utcnow().isoformat(),
            }

            # Upsert calendar integration
            response = (
                supabase_client.table("calendar_integrations")
                .upsert(integration_data, on_conflict="user_id,provider")
                .execute()
            )

            logger.info("Google Calendar connected", user_id=self.user_id)
            return response.data[0]
        except Exception as e:
            logger.error("Failed to connect Google Calendar", error=str(e))
            raise

    async def connect_outlook_calendar(
        self, authorization_code: str
    ) -> Dict[str, Any]:
        """
        Connect user's Outlook Calendar using OAuth code
        """
        try:
            # TODO: Implement full OAuth flow with MSAL
            # This is a placeholder for the OAuth process

            integration_data = {
                "user_id": self.user_id,
                "provider": "outlook",
                "access_token": "placeholder_access_token",
                "refresh_token": "placeholder_refresh_token",
                "token_expires_at": (
                    datetime.utcnow() + timedelta(hours=1)
                ).isoformat(),
                "sync_enabled": True,
                "created_at": datetime.utcnow().isoformat(),
                "updated_at": datetime.utcnow().isoformat(),
            }

            response = (
                supabase_client.table("calendar_integrations")
                .upsert(integration_data, on_conflict="user_id,provider")
                .execute()
            )

            logger.info("Outlook Calendar connected", user_id=self.user_id)
            return response.data[0]
        except Exception as e:
            logger.error("Failed to connect Outlook Calendar", error=str(e))
            raise

    async def sync_tasks_to_calendar(
        self, provider: str, tasks: List[Dict]
    ) -> Dict[str, Any]:
        """
        Sync scheduled tasks to external calendar
        """
        try:
            # Get integration
            integration_response = (
                supabase_client.table("calendar_integrations")
                .select("*")
                .eq("user_id", self.user_id)
                .eq("provider", provider)
                .execute()
            )

            if not integration_response.data:
                raise ValueError(f"No {provider} calendar integration found")

            integration = integration_response.data[0]

            if not integration.get("sync_enabled"):
                raise ValueError(f"{provider} calendar sync is disabled")

            # Sync tasks based on provider
            if provider == "google":
                result = await self._sync_to_google_calendar(integration, tasks)
            elif provider == "outlook":
                result = await self._sync_to_outlook_calendar(integration, tasks)
            else:
                raise ValueError(f"Unsupported calendar provider: {provider}")

            # Update last sync time
            supabase_client.table("calendar_integrations").update(
                {"last_sync_at": datetime.utcnow().isoformat()}
            ).eq("id", integration["id"]).execute()

            logger.info(
                "Tasks synced to calendar",
                user_id=self.user_id,
                provider=provider,
                task_count=len(tasks),
            )

            return result
        except Exception as e:
            logger.error("Failed to sync tasks to calendar", error=str(e))
            raise

    async def _sync_to_google_calendar(
        self, integration: Dict, tasks: List[Dict]
    ) -> Dict[str, Any]:
        """
        Sync tasks to Google Calendar
        """
        # TODO: Implement actual Google Calendar API integration
        # This is a placeholder
        return {
            "provider": "google",
            "synced_count": len(tasks),
            "status": "success",
        }

    async def _sync_to_outlook_calendar(
        self, integration: Dict, tasks: List[Dict]
    ) -> Dict[str, Any]:
        """
        Sync tasks to Outlook Calendar
        """
        # TODO: Implement actual Outlook Calendar API integration
        # This is a placeholder
        return {
            "provider": "outlook",
            "synced_count": len(tasks),
            "status": "success",
        }

    async def disconnect_calendar(self, provider: str) -> None:
        """
        Disconnect a calendar integration
        """
        try:
            supabase_client.table("calendar_integrations").delete().eq(
                "user_id", self.user_id
            ).eq("provider", provider).execute()

            logger.info(
                "Calendar disconnected", user_id=self.user_id, provider=provider
            )
        except Exception as e:
            logger.error("Failed to disconnect calendar", error=str(e))
            raise

    async def get_calendar_integrations(self) -> List[Dict]:
        """
        Get user's calendar integrations
        """
        try:
            response = (
                supabase_client.table("calendar_integrations")
                .select("id, provider, sync_enabled, last_sync_at, created_at")
                .eq("user_id", self.user_id)
                .execute()
            )

            return response.data
        except Exception as e:
            logger.error("Failed to get calendar integrations", error=str(e))
            return []
