backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database & Supabase
supabase==2.3.0
postgrest==0.13.2
sqlalchemy==2.0.25
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/LLM Integration
anthropic==0.18.0
openai==1.10.0
tiktoken==0.6.0

# Async & HTTP
httpx>=0.24.0,<0.25.0
aiofiles==23.2.1
python-multipart==0.0.6

# Calendar Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
msal==1.26.0

# Image & Audio Processing
pillow==10.2.0
opencv-python-headless==4.9.0.80
SpeechRecognition==3.10.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1

# Testing
pytest>=7.0.0,<8
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Monitoring & Logging
structlog==24.1.0
sentry-sdk==1.40.0
